move utils to test.utils
